Add UX Features and Fix Issues
Added hover feature to table used by the app.

Fixed issue where if user exists browser tab, logs off in another tab
on the same browser, or clears cookies/cache from their, authentication
redux state is reset and user will have to log back in.

Added a Link on the Navigation Bar to access the background jobs
and queue service from the management system instead of previously
having to type it in the URL.